Updated album-info file names
