+++ public/js/album-info.js
@@ -1,29 +1,80 @@
-const fetch = require('node-fetch');
-//artist album fetch get artist id from previous fetch
-const deezerAlbumUrl = "https://api.deezer.com/artist/27/albums"
+let btn = document.getElementById('favBtn');
+let url = window.location.pathname;
+let album_id = url.replace(/^\D+/g, '')
 
-let album = async () => {
-    const response = await fetch(deezerAlbumUrl);
-    if (!response.ok) {
-        console.log('error');
+let getFav = async (req, res) => {
+    const response = await fetch('/api/albumfavorites', {
+        method: 'GET',
+        headers: {
+            'Accept': 'application.json',
+            'Content-Type': 'application/json',
+        },
+    });
+    const getData = await response.json();
+    for (var i = 0; i < getData.length; i++) {
+        if (getData[i].album_id == album_id) {
+            btn.classList.remove('deactive');
+            btn.classList.add('active');
+        }
     }
-    return (response.json());
-};
+}
 
-album().then(res => {
-    console.log(res);
+// give Favorite Button red(active) color
+let favBtnActive = function () {
+    btn.classList.remove('deactive');
+    btn.classList.add('active');
+}
+
+// give Favorite Button grey(deactive) color
+let favBtnDeactive = function () {
+    btn.classList.remove('active');
+    btn.classList.add('deactive');
+}
+// Post favorite
+let postDB = async (req, res) => {
+    let url = window.location.pathname;
+    let album_id = url.replace(/^\D+/g, '')
+    let album_name = document.getElementById('title').textContent;
+    const response = await fetch('/api/albumfavorites', {
+        method: 'POST',
+        headers: {
+            'Accept': 'application.json',
+            'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+            album_id: album_id,
+            album_name: album_name,
+        })
+    });
+    const postData = await response.json();
+    console.log(postData);
+}
+// Delete favorite
+let deleteDB = async (req, res) => {
+    let url = window.location.pathname;
+    let album_id = url.replace(/^\D+/g, '')
+    const response = await fetch('/api/albumfavorites', {
+        method: 'DELETE',
+        headers: {
+            'Accept': 'application.json',
+            'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+            album_id,
+        })
+    });
+    const postData = await response.json();
+    console.log(postData);
+}
+
+btn.addEventListener('click', function () {
+    if (btn.classList.contains('active')) {
+        favBtnDeactive();
+        deleteDB();
+    } else {
+        favBtnActive();
+        postDB();
+    }
 });
-// get from album id from previous fetch 
-const deezerTrackListUrl = "https://api.deezer.com/album/302127"
 
-let trackList = async () => {
-    const response = await fetch(deezerTrackListUrl);
-    if (!response.ok) {
-        console.log('error');
-    }
-    return (response.json());
-};
-
-trackList().then(res => {
-    console.log(res);
-})
+window.onload = getFav();
