+++ views/album-info.handlebars
@@ -77,5 +77,5 @@
 </section>
 
 {{#if loggedIn}}
-<script src="/js/tracklist.js"></script>
+<script src="/js/album-info.js"></script>
 {{/if}}
